+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `--torchrun` flag to `gantry run`, which is a shortcut for configuring your experiment and `torchrun`
+  to run your command with all GPUs across all replicas.
 - Added `--text` option to `gantry list` command for filtering by name or description.
 - Added `client` parameter to `api.update_workload_description()` for providing an existing Beaker client,
   which avoids creating one each time the function is called.